backend/requirements.txt
fastapi>=0.103.1
uvicorn[standard]>=0.23.2
requests>=2.31.0
google-genai
riotwatcher>=3.2.3
pydantic>=2.4.2
python-dotenv>=1.0.0
